fix(transfer): drop double errno suffix in sender link_stat/opendir errors
The [sender]-side flist walk and send-open failure paths printed the
io::Error Display (which already ends in " (os error N)") and then
appended " (N)" again, yielding "... (os error 2) (2)". Upstream rsyserr
prints only "<strerror> (<errno>)".

Hoist the engine upstream_io_error() helper to a shared pub fn and reuse
it across the engine link_stat error, the core client I/O errors, and the
seven transfer-crate eprintln sites (walk.rs link_stat/make_file/opendir/
readdir, protocol_io.rs send_files-open), removing the local copy in core.
